+++ app/src/main/res/raw/dpkg.bin
@@ -1,0 +1,66 @@
+#!/data/data/vn.vhn.vsc/files/usr/bin/sh
+
+echo '$ dpkg.bin ' $@ >&2
+
+DPKG_UNPACK=0
+for p in "$@"
+do
+    if [ "$p" = "--unpack" ]; then
+        DPKG_UNPACK=1
+    fi
+done
+
+if [ "$DPKG_UNPACK" = "1" ]; then
+    for p in "$@"; do
+        if [ -f "$p" ]; then
+            echo "> DPKG Process $p" >&2
+            sed -i'' -e 's#/data/data/com.termux#/data/data/vn.vhn.vsc#' $p
+        fi
+    done
+fi
+
+if [ "$DPKG_UNPACK" = "1" ]; then
+	DPKG_INSTALL_ROOT=/data/data/vn.vhn.vsc/files/tmp/vh-dpkg.bin-install
+	rm -rf $DPKG_INSTALL_ROOT || true
+	mkdir -p $DPKG_INSTALL_ROOT
+	dpkg.bin-orig --instdir=$DPKG_INSTALL_ROOT "$@" || exit $?
+	find $DPKG_INSTALL_ROOT/data/data/com.termux/files |  while read f; do
+		g=$(echo $f | sed -e 's#/data/data/com.termux#/data/data/vn.vhn.vsc#')
+		#echo "Work $f $g" >&2
+		if [ -L "$f" ]; then
+			target=$(readlink $f)
+			target_replaced=$(echo $target | sed -e 's#/data/data/com.termux#/data/data/vn.vhn.vsc#')
+			rm -f $g
+			ln -s "$target_replaced" "$g"
+		elif [ -d "$f" ]; then
+			mkdir -p "$g"
+			chmod "--reference=$g" "$f"
+		else
+			cp -p $f $g
+		fi
+	done
+	rm -rf $DPKG_INSTALL_ROOT/data/data/com.termux
+	find $DPKG_INSTALL_ROOT |  while read f; do
+		g=$(echo $f | sed -e "s#$DPKG_INSTALL_ROOT/#/#")
+		#echo "Work2 $f $g" >&2
+		if [ -L "$f" ]; then
+			target=$(readlink $f)
+			rm -f $g
+			ln -s "$target" "$g"
+		elif [ -d "$f" ]; then
+			mkdir -p "$g"
+			chmod "--reference=$g" "$f"
+		else
+			cp -p $f $g
+			sed -i'' -e 's#/data/data/com.termux#/data/data/vn.vhn.vsc#' $g
+		fi
+	done
+	find /data/data/vn.vhn.vsc/files/usr/var/lib/dpkg |  while read f; do
+	  if [ -f "$f" ]; then
+			sed -i'' -e 's#/data/data/com.termux#/data/data/vn.vhn.vsc#' $f
+    fi
+  done
+else
+	dpkg.bin-orig "$@" || exit $?
+fi
+echo 'Finished $ dpkg.bin ' $@ >&2
